+++ packages/db/src/domains/navigation/world/edges.ts
@@ -1,6 +1,7 @@
 import type { ResourceName } from '../../resource/index.js';
 
 import { inBuckets } from './in-buckets.js';
+import { EAST_BLUE_EDGES } from './islands/east-blue/edges.js';
 import type { Island, Sea } from './zones.js';
 
 type TravelRequirement = { kind: 'item'; name: ResourceName };
@@ -24,46 +25,8 @@ export type Edge = {
   modifiers?: Array<TravelModifier>;
 };
 
-const EAST_BLUE_TRAVEL_DURATION = inBuckets('2h');
-
 export const WORLD_EDGES: Array<Edge> = [
-  { from: 'satsuruzo', to: 'yotsuba', via: 'sea_east_blue', baseDurationBuckets: EAST_BLUE_TRAVEL_DURATION },
-  { from: 'satsuruzo', to: 'nagagutsu', via: 'sea_east_blue', baseDurationBuckets: EAST_BLUE_TRAVEL_DURATION },
-  { from: 'dawn', to: 'goat', via: 'sea_east_blue', baseDurationBuckets: inBuckets('1h') },
-  { from: 'dawn', to: 'yotsuba', via: 'sea_east_blue', baseDurationBuckets: EAST_BLUE_TRAVEL_DURATION },
-  { from: 'goat', to: 'sixis', via: 'sea_east_blue', baseDurationBuckets: inBuckets('1h') },
-  { from: 'goat', to: 'nagagutsu', via: 'sea_east_blue', baseDurationBuckets: EAST_BLUE_TRAVEL_DURATION },
-  { from: 'goat', to: 'organ', via: 'sea_east_blue', baseDurationBuckets: EAST_BLUE_TRAVEL_DURATION },
-  { from: 'goat', to: 'kumate', via: 'sea_east_blue', baseDurationBuckets: EAST_BLUE_TRAVEL_DURATION },
-  { from: 'goat', to: 'yotsuba', via: 'sea_east_blue', baseDurationBuckets: EAST_BLUE_TRAVEL_DURATION },
-  { from: 'yotsuba', to: 'organ', via: 'sea_east_blue', baseDurationBuckets: EAST_BLUE_TRAVEL_DURATION },
-  { from: 'yotsuba', to: 'nagagutsu', via: 'sea_east_blue', baseDurationBuckets: EAST_BLUE_TRAVEL_DURATION },
-  { from: 'yotsuba', to: 'kumate', via: 'sea_east_blue', baseDurationBuckets: EAST_BLUE_TRAVEL_DURATION },
-  { from: 'mirrorball', to: 'nagagutsu', via: 'sea_east_blue', baseDurationBuckets: inBuckets('1h') },
-  { from: 'mirrorball', to: 'organ', via: 'sea_east_blue', baseDurationBuckets: EAST_BLUE_TRAVEL_DURATION },
-  { from: 'mirrorball', to: 'gecko', via: 'sea_east_blue', baseDurationBuckets: EAST_BLUE_TRAVEL_DURATION },
-  { from: 'nagagutsu', to: 'organ', via: 'sea_east_blue', baseDurationBuckets: inBuckets('1h') },
-  { from: 'nagagutsu', to: 'gecko', via: 'sea_east_blue', baseDurationBuckets: EAST_BLUE_TRAVEL_DURATION },
-  { from: 'organ', to: 'gecko', via: 'sea_east_blue', baseDurationBuckets: EAST_BLUE_TRAVEL_DURATION },
-  { from: 'organ', to: 'kumate', via: 'sea_east_blue', baseDurationBuckets: EAST_BLUE_TRAVEL_DURATION },
-  { from: 'organ', to: 'rare_animal', via: 'sea_east_blue', baseDurationBuckets: EAST_BLUE_TRAVEL_DURATION },
-  { from: 'rare_animal', to: 'gecko', via: 'sea_east_blue', baseDurationBuckets: EAST_BLUE_TRAVEL_DURATION },
-  { from: 'rare_animal', to: 'kumate', via: 'sea_east_blue', baseDurationBuckets: inBuckets('1h') },
-  { from: 'rare_animal', to: 'tequila_wolf', via: 'sea_east_blue', baseDurationBuckets: EAST_BLUE_TRAVEL_DURATION },
-  { from: 'kumate', to: 'sixis', via: 'sea_east_blue', baseDurationBuckets: EAST_BLUE_TRAVEL_DURATION },
-  { from: 'kumate', to: 'tequila_wolf', via: 'sea_east_blue', baseDurationBuckets: EAST_BLUE_TRAVEL_DURATION },
-  { from: 'sixis', to: 'tequila_wolf', via: 'sea_east_blue', baseDurationBuckets: inBuckets('1h') },
-  { from: 'gecko', to: 'baratie', via: 'sea_east_blue', baseDurationBuckets: EAST_BLUE_TRAVEL_DURATION },
-  { from: 'baratie', to: 'conomi', via: 'sea_east_blue', baseDurationBuckets: EAST_BLUE_TRAVEL_DURATION },
-  { from: 'baratie', to: 'oykot', via: 'sea_east_blue', baseDurationBuckets: EAST_BLUE_TRAVEL_DURATION },
-  { from: 'baratie', to: 'pole_star', via: 'sea_east_blue', baseDurationBuckets: EAST_BLUE_TRAVEL_DURATION },
-  { from: 'conomi', to: 'cozia', via: 'sea_east_blue', baseDurationBuckets: inBuckets('1h') },
-  { from: 'conomi', to: 'frauce', via: 'sea_east_blue', baseDurationBuckets: EAST_BLUE_TRAVEL_DURATION },
-  { from: 'conomi', to: 'oykot', via: 'sea_east_blue', baseDurationBuckets: EAST_BLUE_TRAVEL_DURATION },
-  { from: 'conomi', to: 'pole_star', via: 'sea_east_blue', baseDurationBuckets: EAST_BLUE_TRAVEL_DURATION },
-  { from: 'cozia', to: 'frauce', via: 'sea_east_blue', baseDurationBuckets: inBuckets('1h') },
-  { from: 'oykot', to: 'pole_star', via: 'sea_east_blue', baseDurationBuckets: inBuckets('1h') },
-  { from: 'pole_star', to: 'reverse_mountain', via: 'sea_east_blue', baseDurationBuckets: 8 },
+  ...EAST_BLUE_EDGES,
   {
     from: 'reverse_mountain',
     to: 'whisky_peak',
