+++ packages/db/src/domains/navigation/world/islands/east-blue/mirrorball.ts
@@ -2,9 +2,9 @@ import { defineIsland } from '../define-island.js';
 
 export const MIRRORBALL = defineIsland({
   key: 'mirrorball',
-  name: 'Mirrorball Island',
+  name: 'Île Mirrorball',
   entrySubZone: 'mirrorball_coast',
   subZones: {
-    mirrorball_coast: 'Côte de Mirrorball Island',
+    mirrorball_coast: 'Côte',
   },
 });
