+++ packages/db/src/domains/navigation/world/islands/east-blue/organ.ts
@@ -2,7 +2,7 @@ import { defineIsland } from '../define-island.js';
 
 export const ORGAN = defineIsland({
   key: 'organ',
-  name: 'Archipel Organ',
+  name: 'Archipel Des Orgao',
   entrySubZone: 'orange_town',
   subZones: {
     orange_town: 'Orange Town',
