+++ packages/db/src/domains/navigation/world/islands/east-blue/registry.ts
@@ -1,0 +1,39 @@
+import { BARATIE } from './baratie.js';
+import { CONOMI } from './conomi.js';
+import { COZIA } from './cozia.js';
+import { DAWN } from './dawn.js';
+import { FRAUCE } from './frauce.js';
+import { GECKO } from './gecko.js';
+import { GOAT } from './goat.js';
+import { KUMATE } from './kumate.js';
+import { MIRRORBALL } from './mirrorball.js';
+import { NAGAGUTSU } from './nagagutsu.js';
+import { ORGAN } from './organ.js';
+import { OYKOT } from './oykot.js';
+import { POLE_STAR } from './pole-star.js';
+import { RARE_ANIMAL } from './rare-animal.js';
+import { SATSURUZO } from './satsuruzo.js';
+import { SIXIS } from './sixis.js';
+import { TEQUILA_WOLF } from './tequila-wolf.js';
+import { YOTSUBA } from './yotsuba.js';
+
+export const EAST_BLUE_ISLAND_REGISTRY = [
+  SATSURUZO,
+  DAWN,
+  GOAT,
+  YOTSUBA,
+  MIRRORBALL,
+  NAGAGUTSU,
+  ORGAN,
+  RARE_ANIMAL,
+  KUMATE,
+  SIXIS,
+  TEQUILA_WOLF,
+  GECKO,
+  BARATIE,
+  CONOMI,
+  COZIA,
+  FRAUCE,
+  OYKOT,
+  POLE_STAR,
+];
