+++ packages/db/src/domains/navigation/world/islands/east-blue/sixis.ts
@@ -2,9 +2,9 @@ import { defineIsland } from '../define-island.js';
 
 export const SIXIS = defineIsland({
   key: 'sixis',
-  name: 'Sixis',
+  name: 'Île Sixis',
   entrySubZone: 'sixis_coast',
   subZones: {
-    sixis_coast: 'Côte de Sixis',
+    sixis_coast: 'Côte',
   },
 });
