+++ packages/db/src/domains/navigation/world/islands/east-blue/tequila-wolf.ts
@@ -2,9 +2,9 @@ import { defineIsland } from '../define-island.js';
 
 export const TEQUILA_WOLF = defineIsland({
   key: 'tequila_wolf',
-  name: 'Tequila Wolf',
+  name: 'Île Tequila Wolf',
   entrySubZone: 'tequila_wolf_coast',
   subZones: {
-    tequila_wolf_coast: 'Côte de Tequila Wolf',
+    tequila_wolf_coast: 'Côte',
   },
 });
