+++ packages/db/src/domains/navigation/world/islands/east-blue/yotsuba.ts
@@ -2,7 +2,7 @@ import { defineIsland } from '../define-island.js';
 
 export const YOTSUBA = defineIsland({
   key: 'yotsuba',
-  name: 'Yotsuba Island',
+  name: 'Île Yotsuba',
   entrySubZone: 'shells_town',
   subZones: {
     shells_town: 'Shells Town',
